Refactor: Move service files to src/services/ directory
- Moved 13 service files to src/services/ using git mv to preserve history
- Created __init__.py file in src/services/
- Updated import paths in models.py and test files to use new src.core location
- Organized service modules for better project structure

Files moved:
- ai_product_service.py
- ai_creative_format_service.py
- ai_parsing_comparison.py
- aee_signals.py
- gam_inventory_service.py
- gam_orders_service.py
- policy_check_service.py
- slack_notifier.py
- targeting_capabilities.py
- targeting_dimensions.py
- foundational_formats.py
- default_products.py
- activity_feed.py
