+++ src/core/__init__.py
@@ -1,0 +1,11 @@
+"""
+Core MCP server components for the AdCP Sales Agent.
+
+This module contains the core functionality for the MCP server including:
+- MCP server implementation (main.py)
+- Data models and schemas (schemas.py)
+- Configuration management (config_loader.py, config_loader_v2.py)
+- Security and audit logging (audit_logger.py)
+- Context management (context_manager.py)
+- Validation logic (validation.py, json_validators.py)
+"""
