+++ src/core/context_manager.py
