+++ src/core/validation.py
