+++ src/services/__init__.py
@@ -1,0 +1,7 @@
+"""
+Service modules for the AdCP Sales Agent Server.
+
+This package contains business logic services that handle specific functionality
+like AI-powered product generation, creative format management, targeting capabilities,
+and integration with external services.
+"""
