+++ src/services/ai_product_service.py
