+++ src/services/default_products.py
